extract changeKeys
